[LAB-11]: 페이지네이션 검증 보강
- PageRequest.offset을 Long 반환으로 변경. page와 size 모두 Int인 상태로 곱하면 큰 page 값에서 음수로 오버플로우될 수 있어, DB OFFSET으로 잘못 전달되는 위험 제거
- PageResult 생성자에 require로 page>=0, size>0, totalElements>=0 검증 추가. PageRequest는 검증하던 비대칭 해소이며 size=0으로 들어오면 totalPages 계산이 division-by-zero가 되던 경로 차단
- PageRequest 경계값 성공 케이스(size=1, MAX_SIZE)와 offset 오버플로우 회귀 테스트 추가. 실패 케이스만 있던 상태에서 inclusive 범위 회귀를 잡기 위함
- PageResult 검증 동작을 확인하는 테스트 추가

+++ lab-common/src/main/kotlin/com/crispinlab/common/pagination/PageRequest.kt
@@ -9,8 +9,8 @@ data class PageRequest(
         require(size in 1..MAX_SIZE) { "size must be in 1..$MAX_SIZE, was $size" }
     }
 
-    val offset: Int
-        get() = page * size
+    val offset: Long
+        get() = page.toLong() * size
 
     companion object {
         const val MAX_SIZE: Int = 200
